원거리 무기 탄약량 ui 구현

+++ Source/UntilDawn/Public/Item/ItemObject.h
@@ -68,7 +68,7 @@ public:
 
 	void SetItemQuantity(const uint8 quantity);
 
-	FORCEINLINE uint8 GetItemQuantity() const { return itemInfo.quantity; }
+	FORCEINLINE uint16 GetItemQuantity() const { return itemInfo.quantity; }
 
 	void SetOwnerController(TWeakObjectPtr<APlayerControllerMainMap> controller);
